Implement NativeViewHostAura::InstallClip.
Despite the NOTIMPLEMENTED comment saying that this method is only
needed for performance, it causes bugs with aura windows in side
Bubbles which need to clip to their parent view rather than the parent
window.

This CL is based on https://codereview.chromium.org/30993004/ but strips
out the gravity code and simplifies the windowing code.

Since the window hierarchy of NPAPI plugins has changed due to this
patch, it was also necessary to modify WebContentsViewAura to reflect
this change.

BUG=377378

Review URL: https://codereview.chromium.org/317823002