[PowerProfiler] Power profiler service keeps running even after devtools frontend has been closed.
This is happening as the power handler is not listening to the frontend close events. When the front-end is closed, the service is unaware of this and it keeps collecting the data. With the active frontend close, the devtools power handler should unregister itself with the service.

BUG=386175

Review URL: https://codereview.chromium.org/341853002