Mojo: Always use Now() instead of HighResNow().
This will hopefully fix/reduce flakiness of
MessagePipeDispatcherTest.BasicThreaded on Windows. This is a no-op on
Mac and non-Mac POSIX (where Now() and HighResNow() are the same).

Apparently, the canonical time is "Now". Our current, weird
implementation of PlatformThread::Sleep() for Windows is:

  // When measured with a high resolution clock, Sleep() sometimes returns much
  // too early. We may need to call it repeatedly to get the desired duration.
  TimeTicks end = TimeTicks::Now() + duration;
  TimeTicks now;
  while ((now = TimeTicks::Now()) < end)
    ::Sleep((end - now).InMillisecondsRoundedUp());

Of course, if Now() is *not* the high res timer ... who knows what the
effect of this is versus HighResNow(). (We could end up having slept for
much longer -- as measured by HighResNow() -- than we wanted. This seems
to be the symptom of this particular flakiness.)

On the other hand, who knows what timer the timeouts of
WaitForSingleObject(), etc. are relative to. So the correctness of times
in Chromium on Windows is highly suspect.

R=yzshen@chromium.org
BUG=387513

Review URL: https://codereview.chromium.org/354533003