Fixes "restore windows" behavior for users who have never changed it.
According to Apple's documentation, the plist which records the "reopen
windows..." setting will not exist unless a login item has been changed.  As
the "reopen windows" is the default behavior, if the plist does not exist, it is
safe to assume the user did not change the setting.

Also makes sure that Chrome does not have this behavior when running on
pre-lion Macs.

BUG=373356

Review URL: https://codereview.chromium.org/338863006