Append caller for template errors in GN.
When GN encounters an error in a template, this now appends the location of the template instantiation to the error.

Previously, if you got an error in a template, it would say something is wrong, but without knowing the caller, it could be difficult to diagnose what the problem was. This fixes the problem.

BUG=368882&

Review URL: https://codereview.chromium.org/345333002