Port HistoryService::QueryRedirects{From,To} to CancelableTaskTracker
The callback no longer receive a HistoryService::Handle but instead a
base::CancelableTaskTracker::TaskId that they can use to cancel the
task via the base::CancelableTaskTracker.

Remove HistoryBackend::GetMostRecentRedirects{From,To} that backed
the HistoryBackend::QueryRedirects{From,To} as they now have the same
interface.

Remove the success and url parameters to the QueryRedirectsCallback
as the url can be bound by the caller if required (most of them don't)
and no code path used the success value.

BUG=371818

Review URL: https://codereview.chromium.org/349153006