[Ozone-DRI] Adding support for modifying the console buffer
This buffer is used by default and the kernel will switch to it if the
application controlling the CRTCs crashes without restoring the CRTC
state. For smooth transitions between login screen and logged-in mode we
want to set the console buffer to a custom image/color.

BUG=none
NOTRY=true

Review URL: https://codereview.chromium.org/351043004