Change mojom_bindings_generator rule_name
gyp uses the rule name to name files in the output directory. For that
reason, we should avoid spaces (and some other things, but definitely
spaces) in rule names.

I've moved the rule name to the message, and made the rule_name
something filename-friendly.

Review URL: https://codereview.chromium.org/349293005