[NaCl SDK] Simplify httpd.py
There was a lot of code for handling example-specific delegates, but we don't
need it anymore and it complicates the HTTP server. I've removed it.

In addition, there was a bug in the delegate search that would cause an
infinite loop, so this is actually fixing a real bug too.

BUG=none
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/351963003