Reland "Update the window labels if they change in overview mode."
Original issue: 331643004

Fixed: use of pointer after being freed in test.

Update the window labels if they change in overview mode.

This code adds a function to a WindowObserver that notfies it of a window title
change, allowing to update the label in overview mode dynamically.

BUG=387130
TBR=flackr@chromium.org, sky@chromium.org
TEST=WindowSelectorTest.CreateLabelUnderWindow

Review URL: https://codereview.chromium.org/352363002