Factor out the dependency from autocomplete_input.
This is the first step of componentization of autocomplete_input.
autocomplete_input depends on several directories in chrome/browser,
but all of them are in Parse(). This CL introduces a new interface
AutocompleteInputSchemeChecker for the task in Parse() and removes
the dependencies from autocomplete_input to the directories.

The next CL will move the autocomplete_input files actually.

BUG=386890
R=pkasting@chromium.org, blundell@chromium.org
TBR=dtrainor@chromium.org, mpcomplete@chromium.org, rdsmith@chromium.org, abarth@chromium.org
TEST=build succeeds

Review URL: https://codereview.chromium.org/347963003