[fsp] Fix parsing a URL in case it is of an isolated type.
Before, the parser class was parsing a virtual_path() member of the passed URL,
which in case of an isolated file system url was of a different format.

This CL resolves this issue by relying on path(), not virtual_path(). The first
one is of the same format for both file system url types.

TEST=unit_test: *FileSystemProvider*Parser_IsolatedURL&
BUG=388963

Review URL: https://codereview.chromium.org/339713007