Fix a browser crash which occurs because the RenderWidgetHostViewAura instance attempts to dereference a deleted LegacyRenderWidgetHostHWND pointer.
In the normal case the destruction of the LegacyRenderWidgetHostHWND instance is controlled by the
RenderWidgetHostViewAura class with a call to the LegacyRenderWidgetHostHWND::Destroy method which in turn
destroys the LegacyRenderWidgetHostHWND instance.

However there are cases like in browser_tests where the main window is destroyed via DestroyWindow which in turn
destroys the LegacyRenderWidgetHostHWND instance via its HWND and then the RenderWidgetHostViewAura instance derefs the LegacyRenderWidgetHostHWND object to destroy it etc.

Fix is to notify the RenderWidgetHostViewAura class that the LegacyRenderWidgetHostHWND instance was destroyed.

BUG=389191,388762
R=sky

Review URL: https://codereview.chromium.org/352223006