[Android] Add an error message for java exception crash
We intentionally crash if jni called java code results in exception.
However the CHECK is in native code. So many developers report the
native stack, but not the java stack, which is usually more helpful.

Add a message to the CHECK(false) for this. And also restructure the
code a bit to avoid repeating the message.

BUG=

Review URL: https://codereview.chromium.org/355213002