[NaCl SDK] nacl_io: Add user callback whenever mount() is called.
This can be used to get the dev number of the newly-mounted filesystem. Then
the node at path "/dev/fs/<dev number>" can be used to send messages to this
filesystem.

The JsFs filesystem uses this technique to allow JavaScript to respond to NaCl
filesystem requests.

BUG=none
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/348153003