Fix browser test bugs with delegated rendering on Mac
Add a delay between requesting a snapshot and actually taking the
snapshot. This is the previously-encountered issue where the snapshot
mechanism lags CALayer content updates by an unspecified amount.

Fix places where defined(USE_AURA) was used to check if delegated
rendering is enabled. In particular, fix places where this was used to
determine if extra error should be given for tab capture tests, and some
places where this was used to determine if ImageTransportSurface should
be initialized.

Add a destroyCompositor method to BrowserCompositorViewCocoa. We
don't precisely control when BrowserCompositorViewCocoa are deallocated,
but we need to precisely control when the ui::Compositor is destroyed.

Use a separate BrowserCompositorViewPlaceholderMac to control the
lifetime of the recyclable BrowserCompositorViewCocoa, rather than
using the scheme where an "invalid" BrowserCompositorViewMac is used
as a placeholder.

BUG=314190
NOTRY=True (passed Mac bots)

Review URL: https://codereview.chromium.org/358403003