[reland] Introduce MimeTypeCollector class for easy MIME type resolution.
The previous reverted CL caused a MSAN leak errors because of a circular
reference to FileBrowserPrivateGetFileTasksFunction, which is a ref counted
class. Instance of it was passed to a callback of a MimeTypeCollector, which
was a member of the ref counted function, and inside was storing the callback
with the ref counted function bound.

As a result, the ref counted FileBrowserPrivateGetFileTasksFunction was never
deleted, and hence resulting in leaking of all its members.

To resolve this problem, the callbacks are always called with a PostTask, and
the callback member of the MimeTypeCollector class is cleared as son as
possible, to remove a reference to the ref counted object.

The reverted CL is here:
FileBrowserPrivateGetFileTasksFunction,

Original description:
This is the first patch, part of efforts to move MIME calculation to one place.
With introducing this class, we no longer need to pass MIME types from Files
app back to C++ layer when running tasks.

The next step will be cleaning up apps/launcher.cc in order to use the new
class.

TBR=benwells@chromium.org
TEST=unit_tests: *FileManagerMimeUtilTest*, browser_tests: *FileSystemProvider*Mime*
BUG=388586

Review URL: https://codereview.chromium.org/367853004