Enable bot_update on windows perf bots
While the performance issue on windows aren't fixed (sh.exe still freezes up),
because the update situation on the bots is so abysmal, I think we're willing
to deal with bot_update pain over svn update pain.

One thing I want to point out explicitly is that two particular bots are known
to fail on cache populate due to timeouts, and the plan is to manually populate
the cache on these bots by hand.

BUG=261741,389361

Review URL: https://codereview.chromium.org/366453002