Fix wonky `if` bracing (possible NULL deref) in spelling_menu_observer.cc
This flew by in a presubmit: spelling_menu_observer.cc:312: If/else
bodies with multiple statements require braces  [readability/braces] [4]

This if statement actually needs the braces to be safe. But it's been
like this for 18 months (r173702), so probably doesn't need a merge.

BUG=12804

Review URL: https://codereview.chromium.org/367843004