Fix a DCHECK hit during app list warm start cases.
The DCHECK was added to ensure the UMA collected was consistent with the
assumptions being made in r270615.

The assumptions were wrong, but we don't want to change the data
collected without first correcting those assumptions.

ShowAppListInteractiveTest.ShowAppListFlag is updated to cover the fixed
DCHECK code path.

BUG=386966

Review URL: https://codereview.chromium.org/362933002