Port HistoryService::ScheduleDBTask to CancelableTaskTracker
The HistoryService::ScheduleDBTask sends HistoryDBTasks to HistoryBackend.
The tasks are later processed unless canceled. If the task RunOnDBThread
method returns false, the task is later re-executed. This pattern requires
the use of CancelableTaskTracker::NewTrackedTaskId to get a Callback that
can be used to check if the task has been canceled yet.

Port all the client code and unit tests to pass a CancelableTaskTracker.

Since CancelableTaskTracker is not thread-safe UrlIndexPrivateData cannot
create its own CancelableTaskTracker, so it is provided by the containing
InMemoryUrlIndex instance and cleaned up before resetting the private_data_
pointer.

BUG=371818
TBR=atwilson,tburkard,pkasting,jcivelli

Review URL: https://codereview.chromium.org/352913002