Clear the CertID of a network if no valid cert exists.
The old behavior of the ClientCertResolver was that a non-empty Cert/KeyID is never cleared. However, this lead to other code handling as if the network was correctly configured even if that Cert/KeyID didn't point to a valid cert anymore.
If a cert is available, the behavior will not change: the resolver always writes the id of the best fitting certificate.

The new behavior is, that if no certificate matching the network's cert pattern is found, then the Cert/KeyID is cleared.
Accordingly, this leads to the Enrollment Dialog correctly popping up.

BUG=390914

Review URL: https://codereview.chromium.org/362323004