Draft implementation of controller pairing UI.
Implemented paring UI for controller side. This implementation lacks styling,
transitions, images and real strings. However it is fully functional, and can
be used with FakeControllerPairingController or with real controller when it will be
implented.
For testing purposes "show-controller-pairing-demo" switch was added. If this
switch is set, paring procedure will be emulated during OOBE. Demo
flow will be configured with value of "show-controller-pairing-demo" option,
or default config will be used if no value for the option is given.
For config format see comments in fake_controller_pairing_controller.h.

BUG=375191
TEST=none

Review URL: https://codereview.chromium.org/342263002