Telemetry: Fix calculation of bounding rects for elements in iframes.
We currently use getBoundingClientRect() to compute the bounds of an element in
the page. That method returns the coordinates wrt to its viewport (which could
be an iframe). For synthetic gestures however, we require coordinates wrt to the
top-most window.

This patch fixes the computation by walking up the tree of frames and updating
the bounds for the element by the coordinates of the frames.

Tested on http://www.polymer-project.org/components/paper-elements/demo.html

BUG=391406

Review URL: https://codereview.chromium.org/365113004