Always attempt to load built-in engines when speaking.
This fixes an issue where built-in voices never appear when you have another extension-based engine installed.

Note that, in theory, this might not be enough to fix the issue in all cases. For example, a caller could get the list of voices while the built-in engine is still loading (yet to add its engine speak/stop listeners). For that particular case, we need to send a callback to the client. See https://codereview.chromium.org/361363003/ for a fix/feature.

BUG=391047

Review URL: https://codereview.chromium.org/364923002