Replace the CHECK call in the PatchServiceManagerCalls function which IAT patches DirectWrite with a DCHECK.
The IAT patches fails with AppCompat if the OpenSCManager intercept is enabled. This is based on crashes we have seen in canary.

In any case that is an unsupported configuration as far as Chrome is concerned because we are compatible with all supported operating
systems. If the IAT patching fails dwrite will fallback to the old behavior of connecting to the font cache service.

BUG=391426
TBR=jam

Review URL: https://codereview.chromium.org/378723002