Refactor handling of --show-app-list command line.
AppListService's RecordStartupTimings is currently broken. To completely
fix, it needs to incorporate the Profile* that StartupBrowserCreatorImpl
wants to load, which is not currently passed in.

This change adds AppListService::HandleLaunchCommandLine, abstracting
away the details of initializing, showing, and recording UMA for the app
list. A follow-up will fix RecordStartupTimings.

BUG=382793

Review URL: https://codereview.chromium.org/350883002