base: Add soft memory limit to DiscardableMemoryManager.
This adds a soft memory limit to the discardable memory manager. Only            
allocations that haven't been used in specific amount of time are                
affected by this limit. This allows a client to use a much higher hard           
limit so temporary spikes in required usage can be handled efficiently           
without causing an increase in usage while idle.                                 
                                                                                 
A "reduce memory usage" function is added to the discardable memory                 
interface to allow the client to effectively trigger purging of memory           
that is affected by the soft limit when system reaches what would be             
considered an idle state.                                                        
                                                                                 
BUG=383361,372158
TEST=base_unittests --gtest_filter=DiscardableMemoryManagerTest.ReduceMemoryUsage

Review URL: https://codereview.chromium.org/336273003