Fixed ProfileSyncServiceBookmarkTest on iOS.
This test skips the managed_node because it's not synced. The test
assumed that the managed_node always comes next to the mobile_node but
this is not true on iOS.

This fix changes the test to look for the managed node via
BookmarkClient::CanSyncNode, which is the correct check in the first
place.

BUG=383398

Review URL: https://codereview.chromium.org/376993002