Allow WebContents key handling to supplant extension overrides of the bookmark shortcut
Use the same prioritization for accelerator processing for the bookmark
shortcut when overridden by an extension as for when it is built-in to
the browser. Namely, allow WebContents key processing to take place
before extension accelerator processing.

Two changes are required:

1. Register the keybinding with normal priority rather than high
priority, so that FocusManager::OnKeyEvent does not process the
accelerator before the WebContents.

2. Don't process the accelerator for the extension in the
AcceleratorTarget subclasses. Instead, use the existing browser command
processing in BookmarkCurrentPage which delegates to the extension. This
is necessary so that the extension accelerator processing does not get
invoked while command execution is otherwise blocked in
BrowserView::PreHandleKeyboardEvent.

This change is for Views. Mac will be addressed in a separate CL.

BUG=389340

Review URL: https://codereview.chromium.org/360423002