Force enterprise enrollment flow upon detection of inconsistent state.
At the time LoginDisplayHostImpl decides whether enrollment flow is to be
started, policy hasn't been read yet, so LoginDisplayHostImpl is not in a
position to decide whether recovery is required.  To work around this, upon
policy load on machines requiring recovery, a flag is stored in prefs which
is accessed by LoginDisplayHostImpl early during (next) boot.

Enrollment recovery depends on the availability of the machine id and is
silently skipped when the latter is missing.  There'll be a separate CL to
have session manager provide the machine id on devices that require recovery.

BUG=389481

Review URL: https://codereview.chromium.org/374573006