Raise bot_update timeout to 30 minutes.
This is meant to give us breathing room and not as a fix, we know that the major blocking
issues have been resolved, but we're still dealing with poor I/O performance on windows.

We're still trying to find ways to rectify those situations, but in the meantime
raise the timeout from 10 minutes to 30 minutes so that we can continue working
on this issue while the git rollout moves forward.  I'm comfortable in doing this
because I believe git is making progress, just slowly.  If a 30 minute timeout
can get us through the first hump of the initial checkout, the subsequent incremental
checkouts should be faster.

As always, this should be reverted once we track down the source of the slowness
(be it upgrading disks, or more msys hackery).

BUG=383455,380958

Review URL: https://codereview.chromium.org/381583005