Make sure WebContents passed to chrome_page_zoom::Zoom() have a ZoomController.
Not all WebContents will be zoomed by chrome_page_zoom::Zoom(), but
those that are need to have a ZoomCOntroller. It is assumed that at
the point at which the WebContents is created it should be known if
it will be passed into chrome_page_zoom::Zoom(), so the ZoomController
is created at that point.

BUG=391708

Review URL: https://codereview.chromium.org/373733003