cc: skipping a horizontal sort by checking the value of vertical_density
When vertical_density is 1,
the value of horizontal_density from do_clustering() should be equal to 1.
Thus, it skips the horizontal sort process in this case.

When I checked logs in google.com and www.naver.com, 
this condition (vertical_density equals to 1) is met 
for more than 70% of cases.

BUG=312861

Review URL: https://codereview.chromium.org/379823002