NavigationTiming: set navigationStart for navigations in new tabs.
This patch uses the existing browser_navigation_start field in
FrameMsg_Navigate to set the navigation start time for browser
navigations in new tabs.

For browser navigations that result in cross-site swap, the value is
later overriden with the end of the beforeunload event of the current
document.

BUG=376004

Review URL: https://codereview.chromium.org/339573003