Implement new design for expanded public session pod design
This CL implements the new design for expanded public session pods from
crbug.com/214904#c40. The basic pod is added now. The advanced pod will
be added in a follow-up CL.

The new design requires a different animation when the pod is expanding:
An expanded pane slides out to the side and part of the main pane
disappears. To facilitate this new animation, the CL refactors the HTML
and CSS implementing user pods to be more modular with several panes
that can overlap each other. This will also make it easier to implement
the new standard pod design in crbug.com/361375.

BUG=214904,241790
TEST=Manual, for all pod types, including desktop

Review URL: https://codereview.chromium.org/369893005