Explicitly resize inner viewport container layer on SetViewportSize
SetViewportSize previously relied on UpdateInnerViewportContainerSize to
resize the inner viewport container layer; however, as of r253909, this
method is a no-op if the container layer has masks_to_bounds set (i.e.
non-Android platforms).

This fix explicitly resizes the inner viewport container layer to be the
same size as the newly resized device viewport before calling
UpdateInnerViewportContainerSize. This fixes the situation in the bug
where the viewport size is changed but nothing in the layer tree changes
so no new frame is committed.

BUG=387678

Review URL: https://codereview.chromium.org/381703002