Pepper: Clarify end-of-bitcode-stream behavior.
The PutBytes() method in PnaclTranslateThread is used for two purposes: giving
data to the translator processes and signaling the end of the stream.  The way
it's currently done is a bit confusing; this breaks the method into two
for clarity. This is to make code review easier to read for an upcoming change
which removes FileDownloader from the trusted plugin.

BUG=370556

Review URL: https://codereview.chromium.org/383213005