aura: Do not synthesize a mouse-move event during a drag.
Synthesizing a mouse-move event during a drag can cause an inconsistent sequence
of events to be dispatched (e.g. pressed, dragged, dragged, moved, dragged ...
released). So do not synthesize mouse-move events when a drag is in progress. The
alternate is to synthesize mouse-dragged event instead of a mouse-move event, but
in multi-display/multi-host setting, the wrong dispatcher may end up synthesizing
the event. So avoid doing that.

BUG=389424, 369247, 366096
R=oshima@chromium.org

Review URL: https://codereview.chromium.org/387863003