Files.app: Disable formatting menu for read only volumes.
The purpose of this change is to disable the menu for write-protected
SD cards that anyhow fails to format.

Strictly speaking this is excessive, because there may be a volume with
read-only file system that still is formattable. We however assume that
the practical need for the case is low (see discussion in the BUG.)

BUG=334108

Review URL: https://codereview.chromium.org/386333002