OpenSSL: forward to BoringSSL.
This change will sit in the OpenSSL deps repo. When we switch to BoringSSL,
that patch will update DEPS to point to this revision. This avoids having to
update webrtc and libjingle repos at the same time.

The NaCl build is removed because I believe that all the pointers to it are in
the main Chromium repo and can be switched over. I'd rather get a clear build
error than linking in both OpenSSL and BoringSSL. (Also, forwarding doesn't
appear to work for NaCl targets.)

BUG=393317

https://codereview.chromium.org/383273002/