ui/platform_window: Introduce PlatformWindow for windowing primitives.
Instead of implementing the native windowing primitives for aura, mojo, ozone
etc., have a single implementation in ui/platform_window, and use that in the
rest of the places. To start off, this CL provides an X11 implementation of
the new interface, and use it for mojo native-viewport.

BUG=361137
R=ben@chromium.org, spang@chromium.org

Review URL: https://codereview.chromium.org/391583002